Rewrote TT and added futility pruning

+++ src/board/null_move.rs
@@ -33,6 +33,6 @@ impl Board {
     }
 }
 
-pub fn null_move_reduction(depth: usize) -> usize {
+pub fn null_move_reduction(depth: u16) -> u16 {
     if depth >= 6 { 3 } else { 2 }
 }
