+++ src/tournament/players.rs
@@ -33,10 +33,10 @@ impl MatchPlayers {
     pub fn from_depth(
         white: String,
         black: String,
-        w_depth: usize,
-        w_q_depth: usize,
-        b_depth: usize,
-        b_q_depth: usize,
+        w_depth: u16,
+        w_q_depth: u16,
+        b_depth: u16,
+        b_q_depth: u16,
     ) -> Self {
         let white_limit = SearchLimits::depth(w_depth, w_q_depth);
         let black_limit = SearchLimits::depth(b_depth, b_q_depth);
